crime_scrapper/requirements.txt
newspaper3k
beautifulsoup4
requests
pandas
